requirements.txt
crewai>=0.30.0
crewai-tools>=0.4.0
litellm>=1.0.0
python-dotenv>=1.0.0
